refactor(bin):  merge script for all terminal tests with file IO
Should reduce cognitive overhead.
